Move code into src
