+++ .prettierrc
@@ -1,4 +1,4 @@
 {
-  "trailingComma": "all",
+  "trailingComma": "none",
   "printWidth": 120
 }
