+++ docker-compose.yml
@@ -1,0 +1,12 @@
+version: "3"
+services:
+  mysql:
+    container_name: socious_mysql
+    image: mysql:5.7
+    ports:
+      - "33060:3306"
+    environment:
+      MYSQL_DATABASE: socious
+      MYSQL_USER: admin
+      MYSQL_PASSWORD: secret
+      MYSQL_ROOT_PASSWORD: root
