--- libs/auth/src/AuthConfig.ts
@@ -1,3 +1,0 @@
-export const config = {
-  secret: "EDlguXthn9Tbg2G7I9IoBbrKvNVe4dUQ75P1j2nx" // todo: make this dynamic
-};
