--- libs/auth/src/Strategies/JwtStrategy.ts
@@ -1,20 +1,0 @@
-import { Injectable } from "@nestjs/common";
-import { PassportStrategy } from "@nestjs/passport";
-import { ExtractJwt, Strategy } from "passport-jwt";
-
-import { config } from "../AuthConfig";
-
-@Injectable()
-export class JwtStrategy extends PassportStrategy(Strategy) {
-  constructor() {
-    super({
-      jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
-      ignoreExpiration: false,
-      secretOrKey: config.secret
-    });
-  }
-
-  public async validate(payload: any) {
-    return { id: payload.sub };
-  }
-}
