--- libs/auth/src/decorators/authenticated-user.decorator.ts
@@ -1,6 +1,0 @@
-import { createParamDecorator, ExecutionContext } from "@nestjs/common";
-
-export const AuthenticatedUser = createParamDecorator((data: unknown, ctx: ExecutionContext) => {
-  const request = ctx.switchToHttp().getRequest();
-  return request.user;
-});
