--- libs/auth/src/index.ts
@@ -1,5 +1,0 @@
-export * from "./AuthModule";
-export * from "./decorators/authenticated-user.decorator";
-export * from "./Guards/JwtAuthGuard";
-export * from "./Guards/LocalAuthGuard";
-export * from "./Services/AuthService";
