--- libs/chat/src/chat.module.ts
@@ -1,14 +1,0 @@
-import { Module } from "@nestjs/common";
-import { TypeOrmModule } from "@nestjs/typeorm";
-
-import { ChatClient, ChatService } from "./chat.service";
-import { Chat } from "./models/chat.model";
-import { Message } from "./models/message.model";
-import { UserChat } from "./models/user-chat.model";
-
-@Module({
-  imports: [TypeOrmModule.forFeature([Chat, Message, UserChat])],
-  providers: [ChatService, ChatClient],
-  exports: [ChatService, ChatClient],
-})
-export class ChatModule {}
