--- libs/chat/src/index.ts
@@ -1,2 +1,0 @@
-export * from './chat.module';
-export * from './chat.service';
