--- libs/users/src/UsersModule.ts
@@ -1,12 +1,0 @@
-import { Module } from "@nestjs/common";
-import { TypeOrmModule } from "@nestjs/typeorm";
-
-import { User } from "./Models/User";
-import { UsersService } from "./Services/UsersService";
-
-@Module({
-  imports: [TypeOrmModule.forFeature([User])],
-  providers: [UsersService],
-  exports: [UsersService]
-})
-export class UsersModule {}
