--- libs/users/src/index.ts
@@ -1,3 +1,0 @@
-export * from "./Models/User";
-export * from "./Services/UsersService";
-export * from "./UsersModule";
