+++ src/Chat/Controllers/ChatController.spec.ts
@@ -1,0 +1,19 @@
+import { Test, TestingModule } from "@nestjs/testing";
+
+import { ChatController } from "./ChatController";
+
+describe("ChatController", () => {
+  let controller: ChatController;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [ChatController],
+    }).compile();
+
+    controller = module.get<ChatController>(ChatController);
+  });
+
+  it("should be defined", () => {
+    expect(controller).toBeDefined();
+  });
+});
