+++ src/Chat/Dto/ChatWith.ts
@@ -1,0 +1,3 @@
+export class ChatWithDto {
+  users: number[];
+}
