+++ src/Chat/Dto/SendMessage.ts
@@ -1,0 +1,3 @@
+export class SendMessageDto {
+  text: string;
+}
