+++ src/Chat/Services/ChatService.spec.ts
@@ -1,0 +1,19 @@
+import { Test, TestingModule } from "@nestjs/testing";
+
+import { ChatService } from "./ChatService";
+
+describe("ChatService", () => {
+  let service: ChatService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [ChatService],
+    }).compile();
+
+    service = module.get<ChatService>(ChatService);
+  });
+
+  it("should be defined", () => {
+    expect(service).toBeDefined();
+  });
+});
