+++ src/Identity/Strategies/JwtStrategy.ts
@@ -1,0 +1,20 @@
+import { Injectable } from "@nestjs/common";
+import { PassportStrategy } from "@nestjs/passport";
+import { ExtractJwt, Strategy } from "passport-jwt";
+
+import { config } from "../../Config";
+
+@Injectable()
+export class JwtStrategy extends PassportStrategy(Strategy) {
+  constructor() {
+    super({
+      jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
+      ignoreExpiration: false,
+      secretOrKey: config.auth.secret,
+    });
+  }
+
+  public async validate(payload: any) {
+    return { id: payload.sub };
+  }
+}
