+++ src/Identity/index.ts
@@ -1,0 +1,7 @@
+export * from "./Decorators/Auditor";
+export * from "./Guards/JwtAuthGuard";
+export * from "./Guards/LocalAuthGuard";
+export * from "./IdentityModule";
+export * from "./Models/User";
+export * from "./Services/AuthService";
+export * from "./Services/UsersService";
