--- src/chat/chat-with.dto.ts
@@ -1,3 +1,0 @@
-export class ChatWithDTO {
-  users: number[];
-}
