--- src/chat/send-message.dto.ts
@@ -1,3 +1,0 @@
-export class SendMessageDTO {
-  text: string;
-}
