+++ src/main.ts
@@ -1,21 +1,10 @@
-import { AuthModule } from "@app/auth";
-import { ChatModule } from "@app/chat";
-import { UsersModule } from "@app/users";
 import { Module } from "@nestjs/common";
 import { NestFactory } from "@nestjs/core";
 import { TypeOrmModule } from "@nestjs/typeorm";
 
-import { ChatController } from "./chat/chat.controller";
-import { AuthController } from "./Controllers/AuthController";
-import { UsersController } from "./Controllers/UsersController";
-
-/*
- |--------------------------------------------------------------------------------
- | Configuration
- |--------------------------------------------------------------------------------
- */
-
-const PORT = process.env.PORT ?? 8370;
+import { ChatModule } from "./Chat/ChatModule";
+import { config } from "./Config";
+import { IdentityModule } from "./Identity";
 
 /*
  |--------------------------------------------------------------------------------
@@ -25,20 +14,14 @@ const PORT = process.env.PORT ?? 8370;
 
 @Module({
   imports: [
-    AuthModule,
-    UsersModule,
     ChatModule,
+    IdentityModule,
     TypeOrmModule.forRoot({
       type: "mysql",
-      host: process.env.DB_HOST || "localhost",
-      port: Number(process.env.DB_PORT || 33060),
-      username: "admin",
-      password: "secret",
-      database: "socious",
       autoLoadEntities: true,
+      ...config.mysql,
     }),
   ],
-  controllers: [AuthController, UsersController, ChatController],
 })
 export class AppModule {}
 
@@ -50,7 +33,7 @@ export class AppModule {}
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
-  await app.listen(PORT);
+  await app.listen(config.port);
 }
 
 bootstrap();
