+++ tsconfig.json
@@ -16,17 +16,6 @@
     "noImplicitAny": false,
     "strictBindCallApply": false,
     "forceConsistentCasingInFileNames": false,
-    "noFallthroughCasesInSwitch": false,
-    "paths": {
-      "@app/*": [
-        "libs/*/src"
-      ],
-      "@app/chat": [
-        "libs/chat/src"
-      ],
-      "@app/chat/*": [
-        "libs/chat/src/*"
-      ]
-    }
+    "noFallthroughCasesInSwitch": false
   }
 }
